feat: service layer complete

+++ src/services/userService.ts
@@ -1,4 +1,5 @@
 import { InternalServerError, NotFoundError } from 'restify-errors';
+import IPost from '../interfaces/post.interface';
 import IUser from "../interfaces/user.interface";
 import UserModel from "../models/user.model";
 
@@ -7,18 +8,14 @@ export default class UserService {
 
     public getAll = async (): Promise<IUser[]> => {
         const users = await this.model.getAllUsers();
-        if (!users) {
-            throw new InternalServerError('internal server error');
-        }
+        if (!users) throw new InternalServerError('internal server error');
 
         return users;
     }
 
     public getUserById = async (id: number): Promise<IUser> => {
         const user = await this.model.getUserById(id);
-        if (!user) {
-            throw new NotFoundError('user not found');
-        }
+        if (!user) throw new NotFoundError('user not found');
 
         return user;
     }
@@ -35,6 +32,9 @@ export default class UserService {
     }
 
     public deleteUser = async (id: number): Promise<void> => {
+        const exist = await this.model.getUserById(id);
+        if (!exist) throw new NotFoundError('user not found')
+
         await this.model.deleteUser(id);
     }
 }
